narok.io initial UI
Implement the initial UI
Improve tests to follow DRY
