+++ fe/.eslintignore
@@ -1,1 +1,2 @@
 src/components/google/GoogleIcon.tsx
+src/components/jestUtil.tsx
