--- fe/src/components/Header/Header.scss
@@ -1,24 +1,0 @@
-@import "../../styles/_variables";
-
-header {
-    min-height: 6vh;
-    height: 6vh;
-
-    img {
-        padding: 10px 10px 0 10px;
-    }
-
-    .header__title {
-        display: inline-block;
-        font-size: $font-size-bigger;
-    }
-
-    select {
-        float: right;
-        margin: 10px;
-        padding: 5px;
-        background-color: $background;
-        color: $primary;
-        border: solid 1px $primary;
-    }
-}
