--- fe/src/components/Home/Home.scss
@@ -1,7 +1,0 @@
-@import "../../styles/_variables";
-
-.container {
-    section {
-        padding: 5px;
-    }
-}
