+++ fe/src/components/main/Header/Header.scss
@@ -1,0 +1,32 @@
+@import "../../../styles/_variables";
+
+header {
+    min-height: 8vh;
+    height: 8vh;
+    background-image: linear-gradient(to bottom, $background, $background-light);
+
+    img {
+        padding: 10px 10px 0 50px;
+        float: left;
+        vertical-align: middle;
+    }
+
+    .header__title {
+        display: inline-block;
+        padding-top: 10px;;
+        font-size: $font-size-bigger;
+    }
+
+    .header__subtitle {
+        display: block;
+    }
+
+    select {
+        float: right;
+        margin: 10px;
+        padding: 5px;
+        background-color: $background;
+        color: $primary;
+        border: solid 1px $primary;
+    }
+}
