+++ fe/src/components/main/Header/Header.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
-import * as i18n from '../../services/i18n/i18n';
-import { AppContext } from '../../services/context/context';
-import CMS from '../../services/i18n/cms';
+import * as i18n from '../../../services/i18n/i18n';
+import { AppContext } from '../../../services/context/context';
+import CMS from '../../../services/i18n/cms';
 import './Header.scss';
 
 export const Header = () => {
@@ -18,17 +18,16 @@ export const Header = () => {
   return (
     <header>
       <img
-        src="/images/flame.png"
+        src="/images/demon-logo-small.png"
         title="Narok.io logo"
         alt="Flame"
       />
       <span className="header__title">{website}</span>
-      {' - '}
-      <span className="header__subtitle">{subtitle}</span>
       <select onChange={changeLanguage}>
         <option value={i18n.languages.english}>{english}</option>
         <option value={i18n.languages.french}>{french}</option>
       </select>
+      <span className="header__subtitle">{subtitle}</span>
 
     </header>
   );
