+++ fe/src/components/main/Home/Home.scss
@@ -1,0 +1,34 @@
+@import "../../../styles/_variables";
+
+.container {
+    width: 100%;
+
+    section {
+        width: 75%;
+        min-height: 100%;
+        height: 100%;
+        background-image: url("/images/thorn-vertical-large.png");
+        background-repeat: repeat-y;
+    }
+
+    aside {
+        vertical-align: top;
+        width: 25%;
+    }
+
+    section,
+    aside {
+        display: inline-block;
+    }
+}
+
+@media screen and (max-width: $media-width-medium) {
+    .container {
+        aside,
+        section {
+            width: 100%;
+            height: unset;
+            min-height: unset;
+        }
+    }
+}
