+++ fe/src/components/routes.tsx
@@ -1,0 +1,9 @@
+import * as React from 'react';
+import { Route, Switch } from 'react-router-dom';
+import { BlogPage } from './blog/BlogPage';
+
+export const Routes = () => (
+  <Switch>
+    <Route exact path="/" component={BlogPage} />
+  </Switch>
+);
