+++ fe/src/index.ejs
@@ -1,8 +1,17 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
+    <meta charset="utf-8">
     <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-    <title>Welcome</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta property="og:url" content="https://narok.io" />
+    <meta property="og:type" content="website" />
+    <meta property="og:title" content="Narok.io" />
+    <meta property="og:description" content="Web-Security and hacking..." />
+    <meta property="og:image" content="/images/demon-logo.png" />
+    <link rel="icon" href="/favicon.ico" type="image/x-icon">
+    <title>Narok.io</title>
 </head>
 
 <body>
