+++ fe/src/index.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-import { Home } from './components/Home/Home';
+import { Home } from './components/main/Home/Home';
 import './styles/_main.scss';
 
 ReactDOM.render(
