+++ fe/src/services/i18n/cms.ts
@@ -8,6 +8,7 @@ const CMS = {
   LANGUAGE_FR: 'LANGUAGE_FR',
   WEBSITE_TITLE: 'WEBSITE_TITLE',
   WEBSITE_SUBTITLE: 'WEBSITE_SUBTITLE',
+  COMINGSOON: 'COMINGSOON',
 };
 
 export default CMS;
