+++ fe/src/styles/_main.scss
@@ -22,8 +22,8 @@ h1:first-child {
 
 
 .container {
-    min-height: 91vh;
-    height: 91vh;
+    min-height: 89vh;
+    height: 89vh;
 }
 
 pre {
