+++ fe/src/styles/_variables.scss
@@ -1,17 +1,21 @@
-$primary: #f7c113;
-$primary-dark: #d68306;
-$primary-light: #678edc;
+$primary: #c90f0f;
+$primary-dark: darken($primary, 5%);
+$primary-light: lighten($primary, 5%);
 $secondary: #000;
-$secondary-dark: white;
-$secondary-light: white;
-$background: #000;
-$surface: #f5f5f6;
+$secondary-dark: darken($secondary, 5%);
+$secondary-light: lighten($secondary, 5%);
+$background: rgb(16, 0, 0);
+$background-dark: darken($background, 5%);
+$background-light: lighten($background, 5%);
+$surface: #f5f5f5;
 $link: rgb(116, 12, 12);
 $on-primary: #fff;
-$on-secondary: #000;
+$on-secondary: #454545;
 $on-background: #000;
 $on-surface: #000;
 $on-error: #000;
 $font-size: 16px;
 $font-size-big: 20px;
 $font-size-bigger: 32px;
+$media-width-large: 1024px;
+$media-width-medium: 768px;
